+++ kernel/mm/kmalloc.h
@@ -18,9 +18,9 @@ bool kmalloc_initialized();
 void kfree(void *buffer);
 
 __malloclike __malloc_dealloc(kfree, 1) __alloc_size(1)
-void *kmalloc(uint64_t size);
+void *kmalloc(uint32_t size);
 
 __malloclike __malloc_dealloc(kfree, 1) __alloc_size(1)
-void *kmalloc_size(uint64_t size, uint64_t *size_out);
+void *kmalloc_size(uint32_t size, uint32_t *size_out);
 
-__alloc_size(2) void *krealloc(void *buffer, uint64_t size);
+__alloc_size(2) void *krealloc(void *buffer, uint32_t size);
